+++ core/src/main/java/com/microsoft/applicationinsights/telemetry/MetricTelemetry.java
@@ -43,13 +43,13 @@ public final class MetricTelemetry extends BaseTelemetry<MetricData> {
     /**
      * Envelope Name for this telemetry.
      */
-    private static final String ENVELOPE_NAME = "Metric";
+    public static final String ENVELOPE_NAME = "Metric";
 
 
     /**
      * Base Type for this telemetry.
      */
-    private static final String BASE_TYPE = "MetricData";
+    public static final String BASE_TYPE = "MetricData";
 
     /**
      * Default constructor
@@ -60,6 +60,11 @@ public final class MetricTelemetry extends BaseTelemetry<MetricData> {
         metric = new DataPoint();
         initialize(data.getProperties());
         data.getMetrics().add(metric);
+    }
+
+    @Override
+    public int getVer() {
+        return getData().getVer();
     }
 
     /**
@@ -214,4 +219,9 @@ public final class MetricTelemetry extends BaseTelemetry<MetricData> {
     public String getBaseTypeName() {
         return BASE_TYPE;
     }
+
+    public DataPointType getKind() {
+        return metric.getKind();
+    }
+
 }