+++ core/src/main/java/com/microsoft/applicationinsights/telemetry/RemoteDependencyTelemetry.java
@@ -43,13 +43,13 @@ public final class RemoteDependencyTelemetry extends BaseSampleSourceTelemetry<R
     /**
      * Envelope Name for this telemetry.
      */
-    private static final String ENVELOPE_NAME = "RemoteDependency";
+    public static final String ENVELOPE_NAME = "RemoteDependency";
 
 
     /**
      * Base Type for this telemetry.
      */
-    private static final String BASE_TYPE = "RemoteDependencyData";
+    public static final String BASE_TYPE = "RemoteDependencyData";
 
 
     /**
@@ -83,6 +83,11 @@ public final class RemoteDependencyTelemetry extends BaseSampleSourceTelemetry<R
         this.data.setData(commandName);
         this.data.setDuration(duration);
         this.data.setSuccess(success);
+    }
+
+    @Override
+    public int getVer() {
+        return getData().getVer();
     }
 
     /**
@@ -384,6 +389,10 @@ public final class RemoteDependencyTelemetry extends BaseSampleSourceTelemetry<R
     @Override
     public String getBaseTypeName() {
         return BASE_TYPE;
+    }
+
+    public String getResultCode() {
+        return getData().getResultCode();
     }
 
     public Map<String, Double> getMetrics() {