+++ core/src/test/java/com/microsoft/applicationinsights/internal/config/TelemetryConfigurationFactoryTest.java
@@ -30,6 +30,7 @@ import com.google.common.base.Predicates;
 import com.google.common.collect.Iterables;
 import com.microsoft.applicationinsights.TelemetryConfiguration;
 import com.microsoft.applicationinsights.channel.concrete.inprocess.InProcessTelemetryChannel;
+import com.microsoft.applicationinsights.channel.concrete.nop.NopTelemetryChannel;
 import com.microsoft.applicationinsights.extensibility.ContextInitializer;
 import com.microsoft.applicationinsights.extensibility.TelemetryInitializer;
 import com.microsoft.applicationinsights.extensibility.TelemetryModule;
@@ -226,7 +227,7 @@ public final class TelemetryConfigurationFactoryTest {
         assertThat(mockConfiguration.getTelemetryProcessors(), Matchers.<TelemetryProcessor>hasItem(instanceOf(RequestTelemetryFilter.class)));
         assertThat(mockConfiguration.getTelemetryProcessors(), Matchers.<TelemetryProcessor>hasItem(instanceOf(ValidProcessorsWithSetters.class)));
         assertThat(mockConfiguration.getTelemetryProcessors(), Matchers.<TelemetryProcessor>hasItem(instanceOf(TestProcessorWithoutSetters.class)));
-    }
+            }
 
     @Test
     public void testTelemetryContextInitializers() {
@@ -388,7 +389,7 @@ public final class TelemetryConfigurationFactoryTest {
     public void testEmptyConfiguration() {
         TelemetryConfiguration emptyConfig = TelemetryConfiguration.getActiveWithoutInitializingConfig();
         Assert.assertEquals(null, emptyConfig.getInstrumentationKey());
-        Assert.assertEquals(null, emptyConfig.getChannel());
+        Assert.assertSame(NopTelemetryChannel.instance(), emptyConfig.getChannel());
         Assert.assertEquals(0, emptyConfig.getTelemetryModules().size());
         Assert.assertEquals(false, emptyConfig.isTrackingDisabled());
         Assert.assertEquals(0, emptyConfig.getContextInitializers().size());