+++ core/src/test/java/com/microsoft/applicationinsights/telemetry/BaseTelemetryTest.java
@@ -64,6 +64,11 @@ public final class BaseTelemetryTest {
         }
 
         @Override
+        public int getVer() {
+            return 0;
+        }
+
+        @Override
         protected void additionalSanitize() {
             ++numberOfCallsToAdditionalSanitize;
         }